web/requirements.txt
django
djangorestframework
celery[mongodb]
pymongo
